+++ client/src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ErrorBoundary from './ErrorBoundary';
-import Sidebar from './Sidebar';
+import Sidebar from './components/Sidebar';
 
 export default () => (
   <ErrorBoundary>
