+++ client/src/components/SetContent-Album.jsx
@@ -1,10 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  Wrapper,
-  Title,
-  Username,
-} from './StyledContent';
+import { Wrapper, Title } from './StyledContent';
 import UserDialog from './UserDialog';
 
 const Album = ({ data }) => (
@@ -12,16 +8,10 @@ const Album = ({ data }) => (
     <Wrapper>
       <Title>
         <div>
-          <Username>
-            <a href="/">
-              <span>
-                <UserDialog data={data.user}>
-                  {data.user.fullName}
-                </UserDialog>
-              </span>
-            </a>
-          </Username>
-          <a href="/"><span>{data.title}</span></a>
+          <UserDialog user={data.user} />
+          <a href="/">
+            <span>{data.title}</span>
+          </a>
         </div>
       </Title>
       {(data.type || data.year) && (
@@ -43,12 +33,12 @@ Album.propTypes = {
     userUsername: PropTypes.string,
     type: PropTypes.string,
     year: PropTypes.number,
-    user: PropTypes.object,
-  }),
+    user: PropTypes.object
+  })
 };
 
 Album.defaultProps = {
-  data: {},
+  data: {}
 };
 
 export default Album;
