+++ client/src/components/SongContent.jsx
@@ -1,12 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import {
-  Wrapper,
-  Stats,
-  Title,
-  Username,
-} from './StyledContent';
-import { formatCount } from '../lib';
+import { Wrapper, Stats, Title } from './StyledContent';
+import { formatCount } from '../utils';
 import UserDialog from './UserDialog';
 
 class SongContent extends Component {
@@ -24,15 +19,7 @@ class SongContent extends Component {
         <Wrapper>
           <Title>
             <div>
-              <Username>
-                <a href={`/${track.user.username}`}>
-                  <span>
-                    <UserDialog data={track.user}>
-                      {track.user.fullName}
-                    </UserDialog>
-                  </span>
-                </a>
-              </Username>
+              <UserDialog user={track.user} />
               <a href={`/${track.id}`}>{track.title}</a>
             </div>
           </Title>
@@ -52,7 +39,6 @@ class SongContent extends Component {
               </li>
             )}
             {track.reposts > 0 && (
-
               <li>
                 <a href="/" className="ministats-reposts">
                   <span>{formatCount(track.reposts)}</span>
@@ -86,8 +72,8 @@ SongContent.propTypes = {
     reposts: PropTypes.number,
     albumId: PropTypes.number,
     userUsername: PropTypes.string,
-    user: PropTypes.object,
-  }).isRequired,
+    user: PropTypes.object
+  }).isRequired
 };
 
 export default SongContent;
