+++ client/src/components/StyledContent.jsx
