+++ client/src/components/UserCard.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
 import PropTypes from 'prop-types';
-import { formatCount } from '../lib';
+import { formatCount } from '../utils';
 import UserCardButton from './UserCardButton';
 import {
   Card,
@@ -12,7 +12,7 @@ import {
   ImageContainer,
   DescriptionContainer,
   StatsContainer,
-  LocationContainer,
+  LocationContainer
 } from './UserCardStyledContent';
 
 class UserCard extends Component {
@@ -21,7 +21,7 @@ class UserCard extends Component {
     const { data } = this.props;
     this.state = {
       followers: data.followers,
-      followed: data.followed,
+      followed: data.followed
     };
     this.handleFollow = this.handleFollow.bind(this);
   }
@@ -38,34 +38,43 @@ class UserCard extends Component {
   handleFollow() {
     let { followers } = this.state;
     const { followed } = this.state;
-    const { data: { username } } = this.props;
+    const {
+      data: { username }
+    } = this.props;
     if (followed) {
       followers -= 1;
     } else {
       followers += 1;
     }
 
-    this.setState((state) => ({ followers, followed: !state.followed }), () => {
-      const requestOptions = {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(this.state),
-      };
-      fetch(`/api/sidebar/users/${username}`, requestOptions)
-        .then((response) => response.json());
-    });
+    this.setState(
+      (state) => ({ followers, followed: !state.followed }),
+      () => {
+        const requestOptions = {
+          method: 'POST',
+          headers: { 'Content-Type': 'application/json' },
+          body: JSON.stringify(this.state)
+        };
+        fetch(
+          `/api/sidebar/users/${username}`,
+          requestOptions
+        ).then((response) => response.json());
+      }
+    );
   }
 
   render() {
-    const {
-      data, position, handleFocus, handleLeave,
-    } = this.props;
+    const { data, position, handleFocus, handleLeave } = this.props;
     const { followers, followed } = this.state;
     return (
       <Card
         position={position}
-        onMouseEnter={(e) => { handleFocus(e); }}
-        onMouseLeave={() => { handleLeave(); }}
+        onMouseEnter={(e) => {
+          handleFocus(e);
+        }}
+        onMouseLeave={(e) => {
+          handleLeave(e);
+        }}
       >
         <Arrow />
         <CardContent>
@@ -75,7 +84,9 @@ class UserCard extends Component {
             </ImageContainer>
           </Avatar>
           <DescriptionContainer>
-            <a href="/" title={`Visit ${data.fullName}'s profile`}>{data.fullName}</a>
+            <a href="/" title={`Visit ${data.fullName}'s profile`}>
+              {data.fullName}
+            </a>
           </DescriptionContainer>
           <StatsContainer>
             <ul className="stats-group">
@@ -107,18 +118,19 @@ UserCard.propTypes = {
     fullName: PropTypes.string.isRequired,
     followed: PropTypes.bool,
     followers: PropTypes.number.isRequired,
-    location: PropTypes.string,
+    location: PropTypes.string
   }),
   handleLeave: PropTypes.func.isRequired,
   handleFocus: PropTypes.func.isRequired,
-  position: PropTypes.number.isRequired,
+  position: PropTypes.number.isRequired
 };
 
 UserCard.defaultProps = {
   data: {
-    avatar: 'https://s3-us-west-1.amazonaws.com/kleio.sidebar/icons/blue-linear-gradient-purple-151x151.png',
-    location: '',
-  },
+    avatar:
+      'https://s3-us-west-1.amazonaws.com/kleio.sidebar/icons/blue-linear-gradient-purple-151x151.png',
+    location: ''
+  }
 };
 
 export default UserCard;
