+++ package.json
@@ -25,6 +25,15 @@
     ],
     "testEnvironment": "enzyme"
   },
+  "prettier": {
+    "bracketSpacing": true,
+    "jsxBracketSameLine": false,
+    "printWidth": 80,
+    "semi": true,
+    "singleQuote": true,
+    "trailingComma": "none",
+    "useTabs": false
+  },
   "dependencies": {
     "axios": "^0.19.2",
     "dotenv": "^8.2.0",
@@ -65,6 +74,7 @@
     "jest-environment-enzyme": "^7.1.2",
     "jest-enzyme": "^7.1.2",
     "nodemon": "^2.0.4",
+    "prettier": "^2.1.1",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.10"
   },
